requirements.txt
mtcnn==0.1.1
tensorflow==2.13.0
keras==2.13.1
keras-vggface==0.6
keras-applications==1.0.8
tqdm==4.66.1
numpy==1.24.3
Pillow==10.0.0 
